# internal/adapters/http/openapi.yaml
...
openapi: 3.0.3
info:
  title: situs API
  description: >-
    Lokaler, schreibgeschützter Dienst für EUNIS-Habitattypen: Art →
    Habitattypen (mit Rolle) und Habitattyp → Arten / Syntaxa / Crosswalks,
    aus gepinnten EEA- und ESy-Artefakten, mit deutschen Labels und einem
    Anhang-I-Crosswalk (annex1).
  version: 0.3.0 # x-release-please-version
  license:
    name: MIT
servers:
  # Relativ, absichtlich: Swagger-UI baut die "Try it out"-Anfragen aus diesem
  # Eintrag. Eine absolute URL mit festem Port zeigt am tatsächlich benutzten
  # Port vorbei, sobald der Dienst nicht auf dem dort genannten läuft. Relativ
  # löst gegen die Herkunft der Seite auf und trifft damit immer den Server, der
  # sie ausgeliefert hat.
  - url: /
    description: Derselbe Server, der diese Spezifikation ausliefert
paths:
...
  /v1/info:
    get:
      summary: Dienst-, Build- und Index-Information
      description: >-
        Nennt neben Name und Version, worauf der Index gebaut ist: die
        vorkommenden Konzept-Backbones, die Zahl der Konzepte, das Gebietsschema
        und die Zahl der Gebiete mit Verbreitungsdaten. So kann ein Client vorab
        prüfen, ob seine Konzept-IDs zu diesem Index passen. Beantwortbar sind
        allerdings nur `wcvp:`-IDs: die Batch-Route prüft gegen ein fest
        einkompiliertes Präfix, nicht gegen diese Auskunft — sie macht einen
        Bruch sichtbar, sie verschiebt die Prüfung nicht. Schlägt eine der
        beiden Index-Abfragen fehl, ist die Antwort 500 und nicht ein mit
        Nullen gefülltes `index`-Objekt.
      operationId: info
      responses:
        "200":
          description: Name, Version und Selbstauskunft des Index.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Info"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/habitat-type/{typology}/{code}:
    get:
      summary: Ein Habitattyp mit Arten, Syntaxa und Crosswalks
      description: >-
        Dieselbe Route trägt EUNIS und Anhang I — ein Habitattyp wird immer über
        (typology, code) adressiert. Die Crosswalk-Liste trägt beides: andere
        EUNIS-Fassungen und Anhang-I-Entsprechungen. Sie ist leer, wenn es keine
        gibt; bei Anhang I ist das der Normalfall.
      operationId: habitatType
      parameters:
        - $ref: "#/components/parameters/Typology"
        - $ref: "#/components/parameters/Code"
        - $ref: "#/components/parameters/Lang"
        - $ref: "#/components/parameters/Area"
        - $ref: "#/components/parameters/OnlyInArea"
      responses:
        "200":
          description: Der Habitattyp.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitatTypeDetail"
        "400":
          $ref: "#/components/responses/InvalidQuery"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/habitat-type/{typology}/{code}/species:
    get:
      summary: Artenliste eines Habitattyps, optional nach Rolle gefiltert
      operationId: habitatTypeSpecies
      parameters:
        - $ref: "#/components/parameters/Typology"
        - $ref: "#/components/parameters/Code"
        - name: role
          in: query
          required: false
          description: Filtert auf eine Rolle; ohne Angabe werden alle geliefert.
          schema:
            $ref: "#/components/schemas/Role"
        - $ref: "#/components/parameters/Area"
        - $ref: "#/components/parameters/OnlyInArea"
      responses:
        "200":
          description: Die Arten des Habitattyps.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpeciesEntry"
        "400":
          $ref: "#/components/responses/InvalidQuery"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/species/search:
    get:
      summary: Namen suchen, die dieser Index selbst führt
      description: >-
        Sucht einen Teilstring (case-insensitiv) in den `verbatim_name`, die
        der Index selbst trägt, und liefert die zugehörige Concept-ID mit.
        **Das ist keine Namensauflösung:** kein Fuzzy-Matching, keine
        Synonyme, keine Autorenvarianten, kein Backbone-Wissen — dafür ist
        hostus zuständig. Diese Route beantwortet ausschließlich „welche
        Namen kennt *dieser* Index". `concept_id` ist `null`, wenn der Ingest
        den Namen nicht auflösen konnte; solche Treffer werden mitgeliefert
        und nicht gefiltert, weil sie zur Wahrheit über den Index gehören.
      operationId: speciesSearch
      parameters:
        - name: q
          in: query
          required: true
          description: Teilstring, mindestens ein Zeichen. Leer ist INVALID_QUERY, nicht „alles".
          schema:
            type: string
          example: fagus
        - name: limit
          in: query
          required: false
          description: Höchstzahl der Treffer. Fehlt der Parameter, gilt 20.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Die Treffer, nach Namen sortiert. Leer, wenn nichts passt.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpeciesSearchHit"
        "400":
          $ref: "#/components/responses/InvalidQuery"
  /v1/species/{conceptId}/habitat-types:
    get:
      summary: Habitattypen einer Art (mit Rolle)
      description: >-
        Die Kernfrage der Exkursions-App. Autark: eine Konzept-ID braucht keinen
        Upstream-Dienst.
      operationId: speciesHabitatTypes
      parameters:
        - name: conceptId
          in: path
          required: true
          description: Konzept-ID aus hostus.
          schema:
            type: string
        - $ref: "#/components/parameters/Lang"
        - $ref: "#/components/parameters/Area"
        - $ref: "#/components/parameters/OnlyInArea"
      responses:
        "200":
          description: Die Habitattypen samt Rolle der Art.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabitatTypeRole"
        "400":
          $ref: "#/components/responses/InvalidQuery"
        "404":
          description: >-
            Der Index kennt keine Fakten zu dieser Konzept-ID (NOT_FOUND). Eine
            Konzept-ID existiert hier nur, weil eine Artenzeile sie trägt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/species/{conceptId}/traits:
    get:
      summary: Pflanzenökologische Zeigerwerte einer Art
      description: >-
        Zeigerwerte aller ingestierten Vokabulare (EIVE, Tichý, Midolo) zu
        einer Konzept-ID, optional auf ein Vokabular gefiltert. Autark: eine
        Konzept-ID braucht keinen Upstream-Dienst. Ein Konzept ohne
        Trait-Daten liefert ein leeres Array, kein 404 — "keine Daten" ist
        hier eine normale Antwort.
      operationId: speciesTraits
      parameters:
        - name: conceptId
          in: path
          required: true
          description: Konzept-ID aus hostus.
          schema:
            type: string
        - name: vocab
          in: query
          required: false
          description: >-
            Filtert auf ein Vokabular (z. B. "eive"). Unbekannter Wert (kein
            ingestiertes Vokabular) -> 400 INVALID_QUERY.
          schema:
            type: string
      responses:
        "200":
          description: Die Zeigerwerte, je Vokabular gruppiert, nie vermischt.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TraitSet"
        "400":
          $ref: "#/components/responses/InvalidQuery"
  /v1/species/habitat-types:
    post:
      summary: Habitattypen zu einem Satz Konzept-IDs (Batch)
      description: >-
        Eine ganze Geländeaufnahme in einer Anfrage, und autark: situs löst zur
        Laufzeit keine verbatim Namen auf, der Aufrufer schickt bereits
        aufgelöste Konzept-IDs. Kann der Index eine ID nicht beantworten, wird
        sie mit known=false und einem reason zurückgemeldet, nie verworfen.
      operationId: speciesSetHabitatTypes
      parameters:
        - $ref: "#/components/parameters/Lang"
        - $ref: "#/components/parameters/Area"
        - $ref: "#/components/parameters/OnlyInArea"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [concept_ids]
              properties:
                concept_ids:
                  type: array
                  description: >-
                    Konzept-IDs, wie der Aufrufer sie aufgelöst hat. Mehr als
                    maxItems Einträge ergeben INVALID_QUERY. Doppelte IDs sind
                    erlaubt: intern wird die Index-Arbeit dedupliziert, die
                    Antwort trägt aber einen Eintrag je Eingabe in
                    Eingabereihenfolge, sodass response[i] zu concept_ids[i]
                    gehört — ausnahmslos. Ein leerer Eintrag ist deshalb
                    INVALID_QUERY und wird nicht übersprungen: übersprungen
                    verschöbe er die ganze Liste des Aufrufers um eins.
                  minItems: 1
                  maxItems: 300
                  items:
                    type: string
                    minLength: 1
                    example: "wcvp:concept:123456"
      responses:
        "200":
          description: >-
            Je Eingabe-Konzept-ID ein Eintrag, in Eingabereihenfolge,
            Duplikate eingeschlossen.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConceptResolution"
        "400":
          $ref: "#/components/responses/InvalidQuery"
  /v1/syntaxon/{id}/habitat-types:
    get:
      summary: Habitattypen einer Pflanzengesellschaft
      description: >-
        Zeigt das m:n — dieselbe Gesellschaft kann in mehreren Habitattypen
        vorkommen.
      operationId: syntaxonHabitatTypes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/Lang"
      responses:
        "200":
          description: Die verknüpften Habitattypen.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabitatTypeSummary"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    Typology:
      name: typology
      in: path
      required: true
      description: >-
        Typologie in ihrer Fassung. Ein leeres Segment fällt auf eunis@2021
        zurück (der Pfad selbst verlangt ein Segment, weglassen lässt sich es
        also nicht). Eine unbekannte Typologie ist INVALID_QUERY, kein
        NOT_FOUND.
      schema:
        type: string
        example: eunis@2021
    Code:
      name: code
      in: path
      required: true
      description: Code des Habitattyps innerhalb der Typologie.
      schema:
        type: string
        example: R22
    Lang:
      name: lang
      in: query
      required: false
      description: >-
        Antwortsprache, alternativ per Accept-Language. Lokalisierte Felder
        kommen additiv hinzu — name_en bleibt immer gesetzt. Default ist en.
      schema:
        type: string
        enum: [en, de]
    Area:
      name: area
      in: query
      required: false
      description: >-
        Gebietscode nach WGSRPD Level 3 — das Frontend leitet ihn aus GPS ab, so
        braucht situs keine ISO-Zuordnung. Ist er gesetzt, tragen die Arten- und
        Habitattyp-Einträge zusätzlich in_area. Ein Code, zu dem der Index keine
        Verbreitungsdaten hat, ist INVALID_QUERY und nicht "kommt nicht vor" —
        sonst sähen ein Tippfehler und eine echte Abwesenheit gleich aus.
      schema:
        type: string
        example: GER
    OnlyInArea:
      name: only_in_area
      in: query
      required: false
      description: >-
        Verwirft die sicheren Abwesenheiten (in_area=false). Was nicht beurteilbar
        ist (in_area fehlt), bleibt immer stehen. Ohne area ist der Parameter
        absichtlich wirkungslos statt ein Fehler. Auf dem Batch-Endpunkt markiert
        area nur, verwirft aber keinen Eintrag: dort gehört response[i] zu
        concept_ids[i], ein Eintrag darf also nicht fehlen.
      schema:
        type: boolean
  responses:
    InvalidQuery:
      description: Die Anfrage ist fehlerhaft (INVALID_QUERY).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Es gibt keinen solchen Eintrag (NOT_FOUND).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # Dieselbe Fehlerhülle wie überall. Sie ist nur an /v1/info ausgewiesen,
    # weil dort ein Auslöser benannt und getestet ist (eine fehlschlagende
    # Index-Abfrage); die anderen Routen können INTERNAL_ERROR ebenfalls
    # liefern, aber nur als unspezifizierten Ausfall.
    InternalError:
      description: Ein Fehler auf der Dienstseite (INTERNAL_ERROR).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Role:
      type: string
      enum: [diagnostic, constant, dominant]
    # Fehlende Angaben sind abwesende Felder, nie ein Platzhalter oder eine
    # substituierte Null: level fehlt bei Typologien ohne Hierarchie, priority
    # gibt es nur bei annex1, name_de nur wenn ein Label existiert.
    # name_de ist ein Objekt, nicht ein String mit Geschwisterfeld: ein Client soll
    # value nicht lesen koennen, ohne provenance daneben zu sehen. Manche Labels
    # sind amtlich, andere hat situs selbst uebersetzt — das darf nicht abtrennbar
    # sein.
    GermanLabel:
      type: object
      required: [value, provenance, source]
      properties:
        value:
          type: string
        vernacular:
          type: string
          description: >-
            Etablierter deutscher Begriff. Fehlt, wo keiner denselben Umfang
            traegt wie der Typ — ein leerer Wert waere eine Behauptung.
        provenance:
          type: string
          enum: [official, curated, derived, situs]
          description: >-
            situs bedeutet: situs hat uebersetzt, keine externe Quelle steht
            dahinter.
        source:
          type: string
    HabitatTypeSummary:
      type: object
      required: [typology, code, name_en]
      properties:
        typology:
          type: string
          example: eunis@2021
        code:
          type: string
          example: R22
        level:
          type: integer
        name_en:
          type: string
        name_de:
          $ref: '#/components/schemas/GermanLabel'
        priority:
          type: boolean
    SyntaxonRef:
      type: object
      required: [id, rank, name]
      properties:
        id:
          type: string
        rank:
          type: string
          enum: [class, order, alliance]
        name:
          type: string
        author:
          type: string
          description: >-
            Autorschafts-Zitat, direkt aus FloraVeg.EU übernommen; fehlt, wenn
            kein FloraVeg-Treffer gefunden wurde.
        parent_id:
          type: string
          description: >-
            Verweist auf einen FloraVeg-Klasse-/Ordnungscode (reine
            String-Referenz, kein Fremdschlüssel); fehlt, wenn unbekannt.
    TraitValue:
      type: object
      required: [dim, value]
      properties:
        dim:
          type: string
        value:
          type: number
        niche_width:
          type: number
          description: Fehlt, wenn das Vokabular keine Nischenbreite liefert (Tichý, Midolo).
        n_systems:
          type: integer
          description: Fehlt, wenn das Vokabular keine Systemanzahl liefert (Tichý, Midolo).
    TraitSet:
      type: object
      required: [vocab, vocab_version, values]
      properties:
        vocab:
          type: string
          example: eive
        vocab_version:
          type: string
          example: "1.0"
        values:
          type: array
          items:
            $ref: "#/components/schemas/TraitValue"
    CrosswalkRef:
      description: >-
        Die Gegenseite einer Entsprechung, gesehen vom abgefragten Typ. Der
        Qualifier liest sich immer als "abgefragter Typ <qualifier> dieser Typ".
      type: object
      required: [typology, code, qualifier]
      properties:
        typology:
          type: string
        code:
          type: string
        qualifier:
          type: string
          example: "="
    AggregateSource:
      description: >-
        name is absent: the index only persists the aggregate's concept id,
        never its name.
      type: object
      required: [concept_id]
      properties:
        concept_id:
          type: string
        name:
          type: string
    SpeciesEntry:
      description: >-
        verbatim_name ist immer gesetzt; concept_id fehlt, wenn der Name nicht
        auf ein Konzept auflösbar war — solche Zeilen werden behalten.
        provenance/derived_from erscheinen nur bei einer aus einer Sammelart
        abgeleiteten Zeile.
      type: object
      required: [verbatim_name, role]
      properties:
        concept_id:
          type: string
        verbatim_name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        fidelity:
          type: number
        constancy:
          type: number
        in_area:
          $ref: "#/components/schemas/InArea"
        provenance:
          type: string
          enum: [derived_from_aggregate]
        derived_from:
          $ref: "#/components/schemas/AggregateSource"
    SpeciesSearchHit:
      type: object
      required: [verbatim_name, concept_id]
      properties:
        verbatim_name:
          type: string
          example: Fagus sylvatica
        concept_id:
          type: string
          nullable: true
          description: >-
            `null`, wenn der Ingest diesen Namen nicht auflösen konnte. Das
            Feld ist immer vorhanden — ein fehlendes und ein leeres Feld
            wären für einen Client nicht unterscheidbar.
          example: wcvp:concept:83891
    HabitatTypeDetail:
      allOf:
        - $ref: "#/components/schemas/HabitatTypeSummary"
        - type: object
          required: [species, syntaxa, crosswalks]
          properties:
            species:
              description: >-
                Nach Rolle gruppiert; die drei bekannten Rollen sind immer
                vorhanden, ggf. als leere Liste.
              type: object
              additionalProperties:
                type: array
                items:
                  $ref: "#/components/schemas/SpeciesEntry"
            syntaxa:
              type: array
              items:
                $ref: "#/components/schemas/SyntaxonRef"
            crosswalks:
              type: array
              items:
                $ref: "#/components/schemas/CrosswalkRef"
    HabitatTypeRole:
      allOf:
        - $ref: "#/components/schemas/HabitatTypeSummary"
        - type: object
          required: [role, syntaxa]
          properties:
            role:
              $ref: "#/components/schemas/Role"
            fidelity:
              type: number
            constancy:
              type: number
            syntaxa:
              type: array
              items:
                $ref: "#/components/schemas/SyntaxonRef"
            in_area:
              $ref: "#/components/schemas/InArea"
            provenance:
              type: string
              enum: [derived_from_aggregate]
            derived_from:
              $ref: "#/components/schemas/AggregateSource"
    InArea:
      description: >-
        Dreiwertig: true/false, wenn der Index Verbreitungsdaten zum Konzept
        hat, und abwesend, wenn die Frage nicht beantwortbar ist (keine
        Konzept-ID oder keine Verbreitungszeilen). Ohne ?area= fehlt das Feld
        immer — Abwesenheit ist hier "unbekannt", nie "kommt nicht vor".

        Auf POST /v1/species/habitat-types trägt nur ConceptResolution selbst
        dieses Feld: in den verschachtelten habitat_types-Einträgen fehlt es
        auch mit ?area=, denn die Aussage gilt der Art und wäre dort für jeden
        Habitattyp derselben Art dieselbe Konstante. Auf
        GET /v1/species/{conceptId}/habitat-types, wo es keine Eintragsebene
        darüber gibt, ist habitat_types[].in_area dagegen gesetzt.
      type: boolean
    ConceptResolution:
      description: >-
        Ein Eintrag der Batch-Antwort. Bei known=false ist habitat_types leer und
        reason sagt, warum: unknown_backbone heißt, das ID-Präfix ist nicht
        `wcvp:` — geprüft wird gegen dieses fest einkompilierte Präfix, nicht
        gegen die Selbstauskunft aus /v1/info; unknown_concept heißt, das Präfix
        stimmt, aber der Index kennt keine Habitat-Fakten zu diesem Konzept. Die
        Eingabe wird in beiden Fällen zurückgemeldet, nie verworfen.
      type: object
      required: [concept_id, known, habitat_types]
      properties:
        concept_id:
          type: string
        known:
          type: boolean
        reason:
          type: string
          enum: [unknown_backbone, unknown_concept]
        in_area:
          $ref: "#/components/schemas/InArea"
        habitat_types:
          type: array
          items:
            $ref: "#/components/schemas/HabitatTypeRole"
    Info:
      type: object
      required: [service, version, index]
      properties:
        service:
          type: string
        version:
          type: string
        index:
          $ref: "#/components/schemas/IndexInfo"
    # Die Selbstauskunft des Index. Jede Zahl ist am Index *gemessen*, nichts
    # davon ist konfiguriert: ein Client soll vorab prüfen können, ob seine
    # Konzept-IDs zu diesem Index passen, statt es an leeren Antworten zu
    # merken. Bewusst nicht enthalten: die *Fassung* der Backbone (etwa
    # "wcvp 2026-06-15") — der Ingest schreibt sie heute nicht mit, und eine
    # erfundene Fassung wäre schlimmer als keine.
    IndexInfo:
      type: object
      required: [concept_backbones, species_with_concept, area_scheme, areas_with_data]
      properties:
        concept_backbones:
          type: array
          description: >-
            Die im Index vorkommenden Konzept-ID-Präfixe, sortiert. Erwartet ist
            genau `["wcvp"]` — nur solche IDs beantwortet die Batch-Route, denn
            sie prüft gegen ein fest einkompiliertes Präfix. Etwas anderes hier
            heißt darum nicht "auch diese IDs sind beantwortbar", sondern
            "Index und Binary passen nicht zueinander".
          items:
            type: string
          example: ["wcvp"]
        species_with_concept:
          type: integer
          description: Zahl der verschiedenen Konzept-IDs im Index.
          example: 3135
        area_scheme:
          type: string
          description: Das Vokabular, aus dem `?area=`-Codes stammen.
          example: wgsrpd_l3
        areas_with_data:
          type: integer
          description: >-
            Zahl der verschiedenen Gebietscodes mit Verbreitungszeilen. 0, solange
            kein Verbreitungs-Ingest gelaufen ist — das ist eine wahre Aussage
            über den Index, kein Platzhalter. Es ist eine Anzahl und keine Liste:
            *welche* Codes gültig sind, sagt keine Route.
          example: 366
    # Die einheitliche Fehlerhülle jeder Antwort (writeError) — identisch zu
    # hostus. Codes: INVALID_QUERY, NOT_FOUND, INTERNAL_ERROR. Es gibt kein
    # UPSTREAM_UNAVAILABLE: der Lesepfad ist autark, zur Laufzeit hängt kein
    # Upstream-Dienst daran. Eine Konzept-ID, die der Index nicht beantworten
    # kann, ist ausdrücklich KEIN Fehler: der Batch-Endpunkt meldet sie je
    # Eingabe mit known=false und einem reason zurück, damit eine unbekannte ID
    # nicht die ganze Anfrage verwirft.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string